fix: component name
