+++ src/components/Tab/Tabs.tsx
@@ -1,8 +1,6 @@
 import * as React from "react";
-import { Badge } from "..";
-import { useTheme } from "../../themes";
 import * as Styled from "./styled";
-import Flex from "../Flex";
+import { Tab } from "./internal/Tab";
 import useEventCallback from "../../hooks/useEventCallback";
 
 type TabsProps = {
@@ -14,53 +12,6 @@ type TabsProps = {
   withBadge?: boolean;
   onChange: (event: any, value: any) => void;
 };
-
-type TabProps = {
-  text: string;
-  count?: number;
-  selected: boolean;
-  withBadge: boolean;
-  onChange?: (event: any, value: any) => void;
-  onClick?: (event: any) => void;
-};
-
-const Tab = React.forwardRef<HTMLButtonElement, TabProps>(
-  ({ text, count, withBadge, selected, onChange, onClick }, ref) => {
-    const theme = useTheme();
-    const badgeColor = selected ? "primary" : "secondary";
-
-    const handleClick = (event: React.MouseEvent<HTMLButtonElement>) => {
-      if (!selected && onChange) {
-        onChange(event, text);
-      }
-      if (onClick) {
-        onClick(event);
-      }
-    };
-
-    return (
-      <Styled.Button
-        ref={ref}
-        key={text}
-        tabIndex={selected ? 0 : -1}
-        theme={theme}
-        selected={selected}
-        onClick={handleClick}
-      >
-        <Flex display="flex" justifyContent="space-between" alignItems="center">
-          <Styled.Text selected={selected}>{text} </Styled.Text>
-          {withBadge ? (
-            <Badge color={badgeColor} type="pill" fontWeight="bold">
-              {count}
-            </Badge>
-          ) : (
-            ""
-          )}
-        </Flex>
-      </Styled.Button>
-    );
-  },
-);
 
 const Tabs = React.forwardRef<HTMLDivElement, TabsProps>(
   ({ data, value, withBadge = false, onChange }, ref) => {
