+++ src/components/Tab/index.tsx
@@ -1,1 +1,1 @@
-export { default } from "./Tab";
+export { default } from "./Tabs";
